FIXME: very bad hack.
Somehow, xbuild mistakenly thinks that some of the PCL Facade assemblies such as
System.Threading.Tasks.dll need to be flagged as CopyLocal=true, causing them to
be copied to the output directory.

This will make Mono abort with a TypeLoadException because it fails to find
types from these assemblies.

Quick and ugly fix is to simply remove these assemblies via a post-build hook.