Include Xamarin.WebTests.Framework, but not Xamarin.WebTests.
We want the test framework with its client / server code and request
abstraction, but not the actual tests.

The actual tests in Xamarin.WebTests require a default
ServicePointManager.ServerCertificateValidationCallback to be
installed (which accepts the default server certificate).

We do not want to install this default callback in here because we
are using the per-request callbacks that were introduced in .NET 4.5.