Updated README.md.
Please note that the code in this module is not ready yet, it has NOT BEEN SECURITY AUDITED
AND NO GUARANTEE ABOUT SECURITY OR STABILITY can be made!

You are very welcome to use this module to test the new TLS code, find bugs, problems,
security issues, anything in it - but do not use it in production environment until it is ready.