Add new DebugNewTls and ReleaseNewTls solution configurations for the Facades.
The Mono.Security.Interface Reference PCL will soon be moved into web-tests,
but to be usable in there, the Facade assembly needs to be buildable with Mono 4.2.

When building inside the new-tls module, we use the DebugNewTls / ReleaseNewTls
configurations for the Facades; these will define the 'MONO_FEATURE_NEWTLS' conditional
and enable the APIs that were added in Mono 4.3.1.